+++ src/blog/react-and-next-js.mdx
@@ -1,5 +1,7 @@
 ---
-title: "The state of things, React & Nextjs updates"
-description: "a description"
-date: "12/3/2023"
+title: 'The state of things, React & Nextjs updates'
+description: 'a description'
+date: '12/3/2023'
 ---
+
+asdpofjasi[d fioaklsjdf ioaksjdf pioasjdf p[askdopf kasopdf k[
