+++ src/blog/welcome.mdx
@@ -1,8 +1,6 @@
 ---
-title: "Two Forms of Pre-rendering" description: "For 2023, I am planning to
-post my learnings & opinions as articles for frontend engineering. I plan on
-writing my takes on new technology as well as sharing my thoughts and patterns."
-date: "12/3/2023"
+title: 'Welcome'
+date: '12/3/2023'
 ---
 
 For 2023, I am planning to post my learnings & opinions as articles for frontend
