+++ src/hooks/getSortedPostData.ts
@@ -1,0 +1,38 @@
+import fs from 'fs'
+import path from 'path'
+import matter from 'gray-matter'
+
+const blogDirectory = path.join(process.cwd(), 'src/blog')
+
+export default function getSortedPostsData() {
+  const fileNames = fs.readdirSync(blogDirectory)
+  const allPostsData = fileNames.map((fileName) => {
+    // Remove ".md" from file name to get id
+    const id = fileName.replace(/\.mdx$/, '')
+
+    // Read markdown file as string
+    const fullPath = path.join(blogDirectory, fileName)
+    const fileContents = fs.readFileSync(fullPath, 'utf8')
+    // Use gray-matter to parse the post metadata section
+    const matterResult = matter(fileContents)
+    // Combine the data with the id
+    //
+    console.log(matterResult)
+    return {
+      id,
+      //  Let's just set a null date to make sure we always have a date field to sort by below
+      date: null,
+      url: `blog/${id}`,
+      content: matterResult.content,
+      ...matterResult.data,
+    }
+  })
+
+  return allPostsData.sort((a, b) => {
+    if (a.date < b.date) {
+      return 1
+    } else {
+      return -1
+    }
+  })
+}
