+++ src/hooks/index.ts
@@ -1,2 +1,3 @@
-export * from './theme/useTheme';
-export * from './useLocalStorage';
+export * from './theme/useTheme'
+export * from './useLocalStorage'
+export * from './getSortedPostData'
