+++ src/pages/index.tsx
@@ -1,41 +1,8 @@
-import fs from 'fs'
-import path from 'path'
-import matter from 'gray-matter'
 import Link from '../components/Link'
 import Heading from '../components/Heading'
 import Card from '../components/Card'
 import Article from '../components/Article'
-
-const blogDirectory = path.join(process.cwd(), 'src/blog')
-
-function getSortedPostsData() {
-  const fileNames = fs.readdirSync(blogDirectory)
-  const allPostsData = fileNames.map((fileName) => {
-    // Remove ".md" from file name to get id
-    const id = fileName.replace(/\.mdx$/, '')
-
-    // Read markdown file as string
-    const fullPath = path.join(blogDirectory, fileName)
-    const fileContents = fs.readFileSync(fullPath, 'utf8')
-    // Use gray-matter to parse the post metadata section
-    const matterResult = matter(fileContents)
-    // Combine the data with the id
-    //
-    return {
-      id,
-      date: null,
-      ...matterResult.data,
-    }
-  })
-
-  return allPostsData.sort((a, b) => {
-    if (a.date < b.date) {
-      return 1
-    } else {
-      return -1
-    }
-  })
-}
+import getSortedPostsData from '../hooks/getSortedPostData'
 
 export async function getServerSideProps() {
   const posts = getSortedPostsData()
@@ -55,8 +22,9 @@ export default function Home(props) {
         where I work on the Frontend Platform Team.
       </header>
       {props.posts.map((post) => {
+        console.log(post)
         return (
-          <Card is="a" href={post.id} key={post.title}>
+          <Card is="a" href={post.url} key={post.title}>
             <Article title={post.title}>
               <p>{post.description}</p>
               {post.date}
